requirements.txt
opencv-python
bottle
python-dotenv
click
bs4
lxml